refactor: split trade team card
