+++ src/features/architect/tradeMachine/OutgoingPicksList.jsx
@@ -1,0 +1,48 @@
+import React from 'react';
+import { areSamePick, formatPick } from '@/utils/architect/tradeHelpers';
+
+export const OutgoingPicksList = ({ team, picks, onTogglePick, onEditPick }) => (
+  <div>
+    <h4 className="text-sm text-white/70 mb-1">Outgoing Picks</h4>
+    <div className="space-y-1">
+      {team.picks?.map((p) => {
+        const exists = picks.some((pk) => areSamePick(pk, p));
+        return (
+          <div
+            key={`${p.year}-${p.round}-${p.via || ''}`}
+            className={`flex items-start justify-between px-3 py-1 rounded-md text-xs transition-colors ${
+              exists ? 'bg-purple-800/30' : 'bg-white/10'
+            }`}
+          >
+            <div>
+              <div>{formatPick(p)}</div>
+              {exists && (
+                <div className="text-white/50 mt-1 space-y-1">
+                  <input
+                    className="bg-black/30 p-1 rounded w-full"
+                    value={p.protection || ''}
+                    placeholder="Protection"
+                    onChange={(e) => onEditPick(p, 'protection', e.target.value)}
+                  />
+                  <input
+                    className="bg-black/30 p-1 rounded w-full"
+                    value={p.note || ''}
+                    placeholder="Note"
+                    onChange={(e) => onEditPick(p, 'note', e.target.value)}
+                  />
+                </div>
+              )}
+            </div>
+            <button onClick={() => onTogglePick(p)} className="text-blue-400 hover:underline">
+              {exists ? 'Remove' : 'Trade'}
+            </button>
+          </div>
+        );
+      })}
+      {team.picks?.length === 0 && (
+        <div className="text-xs text-white/40">No picks available</div>
+      )}
+    </div>
+  </div>
+);
+
